sim/benchmarking/requirements.txt
cmdbench==0.1.13
matplotlib==3.7.0
numpy==1.24.2
psutil==5.9.4
py-cpuinfo==9.0.0